# openapi.yaml
openapi: 3.0.3
info:
  title: RHESS - Enterprise Skills Server
  description: >-
    Self-hosted AI agent skills directory. Skills Catalog and Discovery endpoints are unauthenticated. Source management
    (write) endpoints require `Authorization: Bearer <RHESS_ADMIN_TOKEN>`.
  version: 0.1.0
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Value of the RHESS_ADMIN_TOKEN environment variable
...
paths:
...
  /api/v1/skills/:
    get:
      summary: List all skills
      tags:
        - Skills
      description: Returns a paginated list of all indexed skills.
      parameters:
        - schema:
            type: integer
            minimum: 1
            default: 1
          in: query
          name: page
          required: false
        - schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          in: query
          name: per_page
          required: false
        - schema:
            type: string
            enum:
              - name
              - updated_at
            default: name
          in: query
          name: sort
          required: false
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: integer
                          description: Internal numeric skill ID
                        source:
                          type: string
                          description: Source slug
                        sourceLabel:
                          type: string
                          description: Human-readable source label
                        sourceUrl:
                          type: string
                          nullable: true
                          description: Source git repository URL
                        slug:
                          type: string
                          description: Skill slug within its source
                        name:
                          type: string
                          description: Skill display name from frontmatter
                        description:
                          type: string
                          description: Skill description from frontmatter
                        artifactType:
                          type: string
                          enum:
                            - skill-md
                            - archive
                          description: Artifact type
                        digest:
                          type: string
                          description: SHA-256 content digest prefixed with 'sha256:'
                        category:
                          type: string
                          nullable: true
                          description: Skill category (null until spec support is added)
                        allowedTools:
                          type: array
                          items:
                            type: string
                          description: Tools the skill is allowed to use
                        skillPath:
                          type: string
                          description: Relative path to SKILL.md within the source repository
                        frontmatter:
                          type: object
                          additionalProperties: true
                          description: Full parsed frontmatter map (excluding name/description)
                        installCommand:
                          type: string
                          description: npx skills add … command to install this skill
                        lastModified:
                          type: string
                          description: ISO 8601 timestamp of last update
                      required:
                        - id
                        - source
                        - slug
                        - name
                        - description
                        - artifactType
                        - digest
                        - allowedTools
                        - skillPath
                        - installCommand
                        - lastModified
                  meta:
                    type: object
                    properties:
                      total:
                        type: integer
                        description: Total number of indexed skills
                      page:
                        type: integer
                        description: Current page number (1-based)
                      per_page:
                        type: integer
                        description: Results per page
                      total_pages:
                        type: integer
                        description: Total number of pages
                      sort:
                        type: string
                        description: Active sort field
                    required:
                      - total
                      - page
                      - per_page
                      - total_pages
                      - sort
                required:
                  - data
                  - meta
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
...
  /api/v1/sources/:
    get:
      summary: List skill sources
      tags:
        - Sources
      description: Returns all registered skill sources with their skill counts and sync status.
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sources:
                    type: array
                    items:
                      type: object
                      properties:
                        id:
                          type: string
                          description: Source slug (used as public identifier)
                        path:
                          type: string
                          description: Git repository URL (alias for url)
                        label:
                          type: string
                          description: Human-readable display name
                        url:
                          type: string
                          description: Git repository URL
                        lastSynced:
                          type: string
                          nullable: true
                          description: ISO 8601 timestamp of last successful sync
                        skillCount:
                          type: integer
                          description: Number of indexed skills from this source
                        status:
                          type: string
                          enum:
                            - idle
                            - syncing
                            - error
                          description: Current sync status
                required:
                  - sources
    post:
      summary: Register a skill source
      tags:
        - Sources
      description: Clones the given git repository, discovers SKILL.md files, and indexes them.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                path:
                  type: string
                  description: HTTPS or SSH git URL (required if url is not supplied)
                label:
                  type: string
                  description: Display name; derived from URL if omitted
                url:
                  type: string
                  description: Legacy alias for path
                slug:
                  type: string
                  description: Legacy explicit slug; derived from label/path if omitted
      security: &ref_0
        - bearerAuth: []
      responses:
        '201':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  source:
                    type: object
                    properties:
                      id:
                        type: string
                        description: Source slug (used as public identifier)
                      path:
                        type: string
                        description: Git repository URL (alias for url)
                      label:
                        type: string
                        description: Human-readable display name
                      url:
                        type: string
                        description: Git repository URL
                      lastSynced:
                        type: string
                        nullable: true
                        description: ISO 8601 timestamp of last successful sync
                      skillCount:
                        type: integer
                        description: Number of indexed skills from this source
                      status:
                        type: string
                        enum:
                          - idle
                          - syncing
                          - error
                        description: Current sync status
                  syncReport:
                    type: object
                    properties:
                      discovered:
                        type: integer
                      indexed:
                        type: integer
                      failed:
                        type: integer
                      failures:
                        type: array
                        items:
                          type: object
                          properties:
                            path:
                              type: string
                            reason:
                              type: string
                required:
                  - source
                  - syncReport
        '400':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
        '401':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
        '403':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
        '409':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
        '422':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
...
  /api/sync:
    post:
      summary: Re-sync all sources
      tags:
        - Sources
      description: >-
        Re-clones and re-indexes every registered source sequentially. Sources already being synced are skipped.
        Rebuilds the search index when complete.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  synced:
                    type: integer
                    description: Number of sources successfully re-synced
                  count:
                    type: integer
                    description: Total number of indexed skills after the sync
                required:
                  - synced
                  - count
        '401':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
        '403':
          description: Default Response
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: object
                    properties:
                      code:
                        type: string
                      message:
                        type: string
                    required:
                      - code
                      - message
...